[node] Further modularize the server
