+++ node/app/controllers/index.js
@@ -1,0 +1,2 @@
+exports.HUDManager = require('./HUDManager');
+exports.EditorSettings = require('./EditorSettings');
